Migrate from v1

+++ updates/000004_create_user_groups.php
@@ -7,19 +7,17 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('user_groups', function(Blueprint $table)
-        {
-            $table->increments('id');
+        Schema::create('user_groups', function(Blueprint $table) {
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('code')->nullable()->index();
             $table->text('description')->nullable();
             $table->timestamps();
         });
 
-        Schema::create('users_groups', function(Blueprint $table)
-        {
-            $table->integer('user_id')->unsigned();
-            $table->integer('user_group_id')->unsigned();
+        Schema::create('users_groups', function(Blueprint $table) {
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('user_group_id')->unsigned();
             $table->primary(['user_id', 'user_group_id'], 'rainlab_user_group');
         });
     }
